Add files via upload
Attached document contains the Java code for automating task 2 of the Amplience Technical assessment.
Note that in the requirements of the assessment the Public Gists field was not present on the UI and therefore a test could not be included. 
In addition to this the numbers provided as part of the spec for the follower/repo count were incorrect so the test utilises the current numbers listed on the profile, however it is arguable that because this is a fluctuating number it would be more beneficial to simply check the fields are present instead of ensuring the data matches and expected result as this would require more upkeep in the long term.